splitting out journey and records to containers with nested route
